requirements.txt
cricketstats==0.1.6
cvxpy==1.4.2
cvxpy==1.6.0
joblib==1.3.2
joblib==1.4.2
lazypredict==0.2.13
matplotlib==3.8.0
matplotlib==3.8.3
numpy==2.1.3
openai==1.55.1
pandas==2.2.3
plotly==5.22.0
plotly==5.24.1
python-dotenv==1.0.1
scikit_learn==1.3.0
seaborn==0.13.2
streamlit==1.40.1
tqdm==4.66.5
tqdm==4.67.0
xgboost==2.1.3
